pyproject.toml
[tool.poetry]
name = "examkit"
version = "0.1.0"
description = "Production-grade exam preparation toolkit - offline, local-only processing of lecture materials"
authors = ["ExamKit Contributors"]
readme = "README.md"
license = "MIT"
packages = [{include = "examkit"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.1"
faster-whisper = "^0.10.0"
ffmpeg-python = "^0.2.0"
pymupdf = "^1.23.8"
pdfminer-six = "^20221105"
python-pptx = "^0.6.23"
pytesseract = "^0.3.10"
Pillow = "^10.1.0"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
spacy = "^3.7.2"
matplotlib = "^3.8.2"
pandas = "^2.1.4"
numpy = "^1.26.2"
scikit-learn = "^1.3.2"
jinja2 = "^3.1.2"
requests = "^2.31.0"
graphviz = "^0.20.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.8"
mypy = "^1.7.1"
types-pyyaml = "^6.0.12"
types-requests = "^2.31.0"

[tool.poetry.scripts]
examkit = "examkit.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=examkit --cov-report=term-missing"
